Renaming pyqir-jit wheel to pyqir-evaluator
