+++ docs/api-reference/index.md
@@ -7,5 +7,5 @@ maxdepth: 2
 
 parser
 generator
-jit
+evaluator
 ```
