+++ examples/README.md
@@ -12,8 +12,8 @@ Intermediate Representation (QIR)](https://github.com/qir-alliance/qir-spec).
   integrate the QIR toolchain into existing Python-based frontends.
 
 - [**Executing
-  QIR**](https://github.com/qir-alliance/pyqir/tree/main/examples/jit): <br/>
-  The [`pyqir-jit`](https://github.com/qir-alliance/pyqir/tree/main/pyqir-jit)
+  QIR**](https://github.com/qir-alliance/pyqir/tree/main/examples/evaluator): <br/>
+  The [`pyqir-evaluator`](https://github.com/qir-alliance/pyqir/tree/main/pyqir-evaluator)
   package provides an easy way to execute generated QIR. It contains the
   necessary [just-in-time
   compilation](https://en.wikipedia.org/wiki/Just-in-time_compilation)
