+++ examples/evaluator/README.md
@@ -1,6 +1,6 @@
-# Examples: pyqir-jit
+# Examples: pyqir-evaluator
 
-The `pyqir-jit` package provides an easy way to execute generated QIR for the
+The `pyqir-evaluator` package provides an easy way to execute generated QIR for the
 purpose of
 
 1. easily testing and experimenting with QIR code
@@ -12,16 +12,16 @@ compilation](https://en.wikipedia.org/wiki/Just-in-time_compilation)
 infrastructure as well an extensibility mechanism to define what actions to
 perform when a gate is applied in Python.
 
-This folder contains the following examples for how to use the `pyqir-jit`
+This folder contains the following examples for how to use the `pyqir-evaluator`
 package:
 
 - **Bernstein-Vazirani example**: <br/>
 This example shows how to log the executed gate sequence
 for a quantum program compiled to [LLVM bitcode](https://www.llvm.org/docs/BitCodeFormat.html).
 It consists of a [Python
-program](https://github.com/qir-alliance/pyqir/tree/main/examples/jit/bernstein_vazirani.py)
+program](https://github.com/qir-alliance/pyqir/tree/main/examples/evaluator/bernstein_vazirani.py)
   that loads the [compiled
-  bitcode](https://github.com/qir-alliance/pyqir/tree/main/examples/jit/bernstein_vazirani.bc)
+  bitcode](https://github.com/qir-alliance/pyqir/tree/main/examples/evaluator/bernstein_vazirani.bc)
   and then uses the `NonadaptiveJit`, and a custom `GateLogger` to print out a
   simple log of the quantum gates applied during execution.
