+++ pyqir-evaluator/MANIFEST.in
