+++ pyqir-evaluator/pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "pyqir-jit"
+name = "pyqir-evaluator"
 version = "0.2.0a1"
 requires-python = ">=3.6"
