+++ pyqir-evaluator/pyqir/evaluator/__init__.py
@@ -1,0 +1,6 @@
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT License.
+
+from pyqir.evaluator._gatelogger import GateLogger as GateLogger
+from pyqir.evaluator._gateset import GateSet as GateSet
+from pyqir.evaluator._nonadaptivejit import NonadaptiveJit as NonadaptiveJit
