+++ pyqir-evaluator/pyqir/evaluator/_nonadaptivejit.py
@@ -1,8 +1,8 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 
-from pyqir.jit._gateset import GateSet
-from pyqir.jit._native import PyNonadaptiveJit
+from pyqir.evaluator._gateset import GateSet
+from pyqir.evaluator._native import PyNonadaptiveJit
 from typing import List, Optional
