+++ pyqir-evaluator/tests/bell_qir_measure.bc
