--- pyqir-jit/pyqir/jit/__init__.py
@@ -1,6 +1,0 @@
-# Copyright (c) Microsoft Corporation.
-# Licensed under the MIT License.
-
-from pyqir.jit._gatelogger import GateLogger as GateLogger
-from pyqir.jit._gateset import GateSet as GateSet
-from pyqir.jit._nonadaptivejit import NonadaptiveJit as NonadaptiveJit
