+++ pyqir-tests/tox.ini
@@ -4,7 +4,7 @@ isolated_build = true
 [testenv]
 deps =
     --editable=../pyqir-generator
-    --editable=../pyqir-jit
+    --editable=../pyqir-evaluator
     pytest
 
 [testenv:test]
