+++ requirements.txt
@@ -1,3 +1,3 @@
 pyqir-generator
 pyqir-parser
-pyqir-jit
+pyqir-evaluator
